Change to official subgraph. Use tokenId for token id.
It was mentioned in discord that the id field on the subgraph for Art Blocks would
change to be `<contract-id>-<token-id>`. In looking to change this it
was noticed that the project was not using the official subgraph for Art
Blocks.

This PR changes the app to use the official subgraph and updates the
attribute that we use for token id to `tokenId`.